+++ src/app/page.tsx
@@ -1,13 +1,11 @@
 'use client'
 
-import Header from '../components/Header'
+import { Header, Footer, FloatingActionButton } from '../components/feature'
 import Home from '../components/sections/Home'
 import Projects from '../components/sections/Projects'
 import Services from '../components/sections/Services'
 import Experience from '../components/sections/Experience'
 import Contact from '../components/sections/Contact'
-import Footer from '../components/Footer'
-import FloatingActionButton from '../components/FloatingActionButton'
 
 export default function Page() {
   return (